+++ application/source/command.cpp
@@ -177,7 +177,7 @@ void Command::backgroundHandler()
 
 void Command::validateBasicTask()
 {
-    for (int i = 0; i < BasicTask::Bottom<BasicTask::Category>::value; ++i)
+    for (uint8_t i = 0; i < BasicTask::Bottom<BasicTask::Category>::value; ++i)
     {
         if (!program.isUsed(std::next(basicTaskMap.cbegin(), BasicTask::Category(i))->first))
         {
@@ -257,7 +257,7 @@ void Command::performTask() const
 {
     if (!assignedTask.basicTask.empty())
     {
-        for (int i = 0; i < BasicTask::Bottom<BasicTask::Category>::value; ++i)
+        for (uint8_t i = 0; i < BasicTask::Bottom<BasicTask::Category>::value; ++i)
         {
             if (assignedTask.basicTask.primaryBit.test(BasicTask::Category(i)))
             {
@@ -267,7 +267,7 @@ void Command::performTask() const
     }
     else if (!assignedTask.generalTask.empty() && !assignedTask.basicTask.primaryBit.test(BasicTask::Category::help))
     {
-        for (int i = 0; i < GeneralTask::Bottom<GeneralTask::Category>::value; ++i)
+        for (uint8_t i = 0; i < GeneralTask::Bottom<GeneralTask::Category>::value; ++i)
         {
             switch (GeneralTask::Category(i))
             {
